% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{flowcut-package}
\alias{flowcut}
\alias{flowcut-package}
\title{flowcut}
\description{
This package implements the \code{flowcut} method.
}
\author{
\strong{Maintainer}: Sangwon Hyun \email{sangwonh@ucsc.edu}

}
\keyword{internal}