% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gen_1d}
\alias{gen_1d}
\title{Generate 1d data with 2 clusters from a list (|true_model|)
containing true model parameters.}
\usage{
gen_1d(true_model, nt = 1000)
}
\arguments{
\item{true_model}{List containing beta, alpha, mn, prob, numclust.}

\item{nt}{Particles per time point.}
}
\value{
Cytograms (a |ylist| object)
}
\description{
Generate 1d data with 2 clusters from a list (|true_model|)
containing true model parameters.
}