% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_Cbox}
\alias{get_Cbox}
\title{Helper to obtain the censoring limits |Cbox|, which is a 2-colum matrix with columns |bounds.lower| and |bounds.upper|.}
\usage{
get_Cbox(ylist)
}
\arguments{
\item{ylist}{List of d-dimensional cytogram data.}
}
\value{
dimdat by 2 matrix
}
\description{
Helper to obtain the censoring limits |Cbox|, which is a 2-colum matrix with columns |bounds.lower| and |bounds.upper|.
}