% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_posterior_probs}
\alias{get_posterior_probs}
\title{Get the membership probabilities of a new dataset |ynew| given GMM model |gmm_model|.}
\usage{
get_posterior_probs(gmm_model, ynew)
}
\arguments{
\item{gmm_model}{from mclust or me.weighted}

\item{ynew}{new data to get predicted posterior membership for.}
}
\value{
estimated posterior probabilities from an estimated GMM model
}
\description{
Get the membership probabilities of a new dataset |ynew| given GMM model |gmm_model|.
}