% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{impute}
\alias{impute}
\title{Imputing particles to impute the y's based on a model defined by the posterior
mean of the last 500 draws. This is separated out from the Gibbs since the result takes up a lot of memory.
(This is all code in the Gibbs.fast() function.)}
\usage{
impute(ylist, countslist, Cbox, mcres, n.cores = 1)
}
\arguments{
\item{ylist}{Original particles}

\item{countslist}{Biomass list}

\item{Cbox}{Bounding box.}

\item{mcres}{The converted MCMC result, using e.g., |mcmc_res_to_flowmix(res,
last_draws_inds = 1001:1500)|}

\item{n.cores}{number of CPU cores to use}
}
\description{
Imputing particles to impute the y's based on a model defined by the posterior
mean of the last 500 draws. This is separated out from the Gibbs since the result takes up a lot of memory.
(This is all code in the Gibbs.fast() function.)
}