% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_sim_model}
\alias{make_sim_model}
\title{From an original set of model parameters (|true_model|),
generate synthetic 2-cluster 1-dimensional data with equal probabilities.}
\usage{
make_sim_model(
  orig_model,
  isignal,
  shrink_alpha = FALSE,
  coef_is_dense = FALSE
)
}
\arguments{
\item{orig_model}{Original model of class flowmix or flowcut; a list that
contains alpha, beta and TT.}

\item{isignal}{0 to 10, which generates the means.}

\item{shrink_alpha}{If TRUE, "shrink" the alpha coefficients to 40\% their
size. Defaults to FALSE.}

\item{coef_is_dense}{If TRUE, add some Gaussian noise on the non-intercept
coefficients to make them dense (instead of sparse).}
}
\value{
A list with beta, mn, alpha, prob, X, sigma, TT, numclust.
}
\description{
From an original set of model parameters (|true_model|),
generate synthetic 2-cluster 1-dimensional data with equal probabilities.
}