% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{maxdev_to_g_closed_form_1d}
\alias{maxdev_to_g_closed_form_1d}
\title{Closed form for obtaining the g parameter by simulation. Currently only works for dimdat=1}
\usage{
maxdev_to_g_closed_form_1d(X, dimdat, maxdev, prior_prob = 0.99)
}
\arguments{
\item{X}{the design matrix, TT by p, (that doesn't include the intercept)}

\item{dimdat}{the dimension of the cytogram space}

\item{maxdev}{Maximum deviation of cluster means away from its grand mean.}

\item{prior_prob}{prior probability}
}
\value{
the g parameter with desired prior probability on maxdev
}
\description{
Closed form for obtaining the g parameter by simulation. Currently only works for dimdat=1
}