% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{maxdev_to_gg}
\alias{maxdev_to_gg}
\title{X is a (T x p) covariate matrix (that doesn't include the intercept)}
\usage{
maxdev_to_gg(X, dimdat, maxdev, viz = FALSE, n.cores = 1)
}
\arguments{
\item{X}{(T x p) matrix.}

\item{dimdat}{Dimension of cytogram data.}

\item{maxdev}{Maximum deviation of cluster means away from its grand mean.}

\item{viz}{Whether to produce a plot.}

\item{n.cores}{Number of cores.}
}
\description{
X is a (T x p) covariate matrix (that doesn't include the intercept)
}