% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{mcmc_res_to_flowmix}
\alias{mcmc_res_to_flowmix}
\title{Reformatting the results from MCMC to create a flowmix-like object that
contains of the posterior means of the model parametr. This allows you to use}
\usage{
mcmc_res_to_flowmix(res, last_draws_inds = NULL)
}
\arguments{
\item{res}{The result of doing MCMC.}
}
\value{
posterior mean estimates of the cluster coefficients, mean, variance
and probability.
}
\description{
Reformatting the results from MCMC to create a flowmix-like object that
contains of the posterior means of the model parametr. This allows you to use
}