% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{MN2SB}
\alias{MN2SB}
\title{Helper to change stick-breaking vector |pi.sb| to regular probabilities. The
opposite function of SB2MN.}
\usage{
MN2SB(pi.mn)
}
\arguments{
\item{pi.mn}{Probability vector.}
}
\value{
same sized vector of size K.
}
\description{
Helper to change stick-breaking vector |pi.sb| to regular probabilities. The
opposite function of SB2MN.
}