% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{post_process_pie}
\alias{post_process_pie}
\title{Description goes here}
\usage{
post_process_pie(Xp, ga)
}
\arguments{
\item{Xp}{1 vector appended to left of covariate matrix X.}

\item{ga}{document this carefully}
}
\value{
(OO x OO) matrix with each row..
}
\description{
Description goes here
}