% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{preimpute}
\alias{preimpute}
\title{At each time point, we will (1) set aside the censored points, and (2) estimate
a 2-component GMM (2) take the censored points, and "impute" them according to those
mixtures. (After that, we'll (3) fit flowmix on the resulting imputed data.)}
\usage{
preimpute(datobj)
}
\arguments{
\item{datobj}{List with ylist (particles), countslist, X, and Cbox.}
}
\value{
datobj with imputed ylist.
}
\description{
This is only designed for 1-dimensional data with two clusters.
}