% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{printprogress}
\alias{printprogress}
\title{A helper function to print the progress of a simulation. Place directly at
the beginning of the loop, before any computation happens.}
\usage{
printprogress(
  isim,
  nsim,
  type = "simulation",
  lapsetime = NULL,
  lapsetimeunit = "seconds",
  start.time = NULL,
  fill = FALSE,
  beep = FALSE
)
}
\arguments{
\item{isim}{isim.}

\item{nsim}{number of sim.}

\item{type}{type of job you're running. Defaults to "simulation".}

\item{lapsetime}{lapsed time, in seconds (by default).}

\item{lapsetimeunit}{"second".}

\item{start.time}{Start time, usually obtained using \code{Sys.time()}}

\item{fill}{Whether or not to fill the line.}

\item{beep}{Whether to beep when done.}
}
\description{
A helper function to print the progress of a simulation. Place directly at
the beginning of the loop, before any computation happens.
}