% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{progress}
\alias{progress}
\title{Helper function to print progress.}
\usage{
progress(iter, Nmc)
}
\arguments{
\item{iter}{Current iteration.}

\item{Nmc}{Total number of MCMC iterations.}
}
\description{
Helper function to print progress.
}