% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{run.Gibbs.fast}
\alias{run.Gibbs.fast}
\title{Runs a gibbs sampler to sample from the posterior distribution of the flowcut
model.}
\usage{
run.Gibbs.fast(
  ylist,
  countslist,
  X,
  numclust,
  Nmc = 3000,
  Nburn = 500,
  Cbox = NULL,
  censorship.info = NULL,
  user.prior = NULL,
  gg = NULL,
  maxdev = NULL,
  maxdev_prob = 0.99,
  prior_spec.list = NULL,
  verbose = FALSE,
  last.imputed = NULL,
  last.para = NULL,
  modified_g_prior = FALSE,
  n.cores = 1
)
}
\arguments{
\item{ylist}{data.}

\item{countslist}{weights (counts) for data in |ylist|.}

\item{X}{covariates; a (p x T) matrix (TODO: change code so that X is T x p).}

\item{numclust}{NUmber of clusters.}

\item{Nmc}{Number of MCMC iterations}

\item{Nburn}{Number of burn-in iterations.}

\item{Cbox}{Censored box.}

\item{user.prior}{User-supplied prior. Otherwise, prior defaults to ___.}

\item{gg}{Size of Normal prior on the beta parameters. Defaults to NULL, in
which case the |maxdev| is used to calculate |gg|. See maxdev_to_gg() for
details. If a value is provided, |maxdev| and |maxdev_prob| are ignored.}

\item{maxdev}{Maximum deviation of the cluster means.}

\item{maxdev_prob}{The probability of a cluster mean deviating from their
overall mean, over time t=1,.., T. Defaults to 0.99.}

\item{verbose}{Whether to be loud.}

\item{modified_g_prior}{Adding a small constant on the diagonal of X^TX to make it
invertible.}

\item{n.cores}{Number of cores for multiple cores.}

\item{warm.start}{If supplied, restart the MCMC at these values.}
}
\description{
Runs a gibbs sampler to sample from the posterior distribution of the flowcut
model.
}