% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{SB2MN}
\alias{SB2MN}
\title{Helper to change stick-breaking vector |pi.sb| to regular probabilities.}
\usage{
SB2MN(pi.sb)
}
\arguments{
\item{pi.sb}{Stickbreaking probability vector.}
}
\value{
same sized vector of size K.
}
\description{
Helper to change stick-breaking vector |pi.sb| to regular probabilities.
}