% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{split_data}
\alias{split_data}
\title{A helper to split the data into (1) censored and (2) non-censored ("inner") particles.}
\usage{
split_data(datobj)
}
\arguments{
\item{datobj}{List containing ylist, countslist, X, and Cbox.}
}
\value{
List of particles on /and/ inside censor boundary. And the directions
in which they were censored.
}
\description{
A helper to split the data into (1) censored and (2) non-censored ("inner") particles.
}