\name{flowCutData}
\alias{flowCutData}
\docType{data}
\title{A list containing two GvHD flow frames from flowCore }
\description{
   We are taking two data examples from flowCore GvHD dataSets. The first data is GvHD flowFrame object 's9a06', the second is 's5a07'. We are also taking two data examples from open source cytometry data repository, FlowRespository. The third data is FlowRespository FR-FCM-ZZ7E Macrophages + oATP.fcs, the fourth is FR-FCM-ZZVB Specimen_001_B6 LSK.fcs. 
   We used the estimateLogical transformation on the data, which will be subsequently processed by flowCut.
}
\usage{data(flowCutData)}
\author{
  Justin Meskas, Sherrie Wang
}
\keyword{datasets}