% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_Cbox}
\alias{get_Cbox}
\title{Helper to obtain the censoring limits |Cbox|, which is a 2-colum matrix with columns |bounds.lower| and |bounds.upper|.}
\usage{
get_Cbox(ylist)
}
\arguments{
\item{ylist}{Data (list of d-column matrices) including censored particles.}
}
\value{
|dimdat| by 2 matrix.
}
\description{
Helper to obtain the censoring limits |Cbox|, which is a 2-colum matrix with columns |bounds.lower| and |bounds.upper|.
}