% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{maxdev_to_gg}
\alias{maxdev_to_gg}
\title{Obtain the g parameter by simulation.}
\usage{
maxdev_to_gg(
  X,
  dimdat,
  maxdev,
  numclust,
  ggvec,
  Nmc = 10000,
  prior.prob = 0.99,
  viz = FALSE,
  n.cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{X}{the design matrix, TT by p, (that doesn't include the intercept)}

\item{dimdat}{the dimension of the cytogram space}

\item{maxdev}{Maximum deviation of cluster means away from its grand mean.}

\item{numclust}{the number of experts}

\item{ggvec}{the vector of g parameter values to calculate the prior probability by Monte Carlo samples}

\item{Nmc}{Monte Carlo simulation sample size, with default value being 1e4.}

\item{viz}{show the plot of the fitted relationship between the g parameter and the prior probability.}

\item{n.cores}{the number of CPU cores to be used for parallelization}

\item{tX}{is the transpose of X}
}
\value{
the g parameter with desired prior probability on maxdev
}
\description{
Obtain the g parameter by simulation.
}