% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{run.Gibbs.fast}
\alias{run.Gibbs.fast}
\title{Runs a gibbs sampler to sample from the posterior distribution of the flowcut
model.}
\usage{
run.Gibbs.fast(
  ylist,
  countslist,
  X,
  numclust,
  Nmc = 3000,
  Nburn = 500,
  Cbox = NULL,
  user.prior = NULL,
  gg = 1,
  prior_spec.list = NULL,
  verbose = FALSE,
  warm.start = NULL,
  tol = 1/1e+08,
  n.cores = 1
)
}
\arguments{
\item{ylist}{data.}

\item{countslist}{weights (counts) for data in |ylist|.}

\item{X}{covariates; a (p x T) matrix (TODO: change code so that X is T x p).}

\item{numclust}{NUmber of clusters.}

\item{Nmc}{Number of MCMC iterations}

\item{Nburn}{Number of burn-in iterations.}

\item{Cbox}{Censored box.}

\item{user.prior}{User-supplied prior. Otherwise, prior defaults to ___.}

\item{gg}{Size of Normal prior.}

\item{verbose}{Whether to be loud.}

\item{warm.start}{If supplied, restart the MCMC at these values.}

\item{tol}{NOT USED.}

\item{n.cores}{Number of cores for multiple cores.}
}
\description{
Runs a gibbs sampler to sample from the posterior distribution of the flowcut
model.
}